replaced menus and prompts with dicts
